pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "remex"
version = "0.5.0"
description = "Retrieval-validated embedding compression. 4-8x smaller vectors, proven recall."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "Oskar Austegard"}]
dependencies = ["numpy>=1.24", "scipy>=1.10"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]
bench = ["faiss-cpu", "sentence-transformers"]

[tool.setuptools.packages.find]
include = ["remex*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
